+++ src/app/FormView/utils/isDataComplete.ts
@@ -1,6 +1,10 @@
-import { Client } from "../services/addClient.service";
+import { clientData } from "@/utils/clientData"
 
-export const isDataComplete = ({email, dni, telf, nombres, primerApl, segundoApl, countryState, city}:Client) => {
+
+export const isDataComplete = () => {
+    const {email, dni, telf, nombres, primerApl, segundoApl, countryState, city} = clientData
+    
+    console.log({email, dni, telf, nombres, primerApl, segundoApl, countryState, city})
     return email !== '' 
         && dni !== '' 
         && telf !== ''
