+++ src/app/resultsView/ResultsView.tsx
