+++ src/store/StepProgress/StepProgressProvider.tsx
@@ -17,17 +17,6 @@ export const StepProgressProvider: React.FC<IStepProgressProvider> = ({
     currentSubstep: 0,
   })
 
-  const [clientData, setClientData] = useState({
-    email: '',
-    dni: '',
-    telf: '',
-    nombres: '',
-    primerApl: '',
-    segundoApl: '',
-    countryState: '',
-    city: '',
-  })
-
   const setTotalSteps = (totalSteps: number) => {
     const updatedStepProgress = { ...stepProgress, totalSteps }
     setStepProgress(updatedStepProgress)
@@ -51,9 +40,7 @@ export const StepProgressProvider: React.FC<IStepProgressProvider> = ({
         setCurrentStep,
         setTotalSubsteps,
         setCurrentSubstep,
-        setStepProgress,
-        clientData,
-        setClientData,
+        setStepProgress
       }}
     >
       {children}
