+++ src/utils/clientData.ts
@@ -1,0 +1,10 @@
+export let clientData ={
+    email: '',
+    dni: '',
+    telf: '',
+    nombres: '',
+    primerApl: '',
+    segundoApl: '',
+    countryState: '',
+    city: '',
+  }
